requirements.txt
beautifulsoup4==4.8.2
Django==2.2
IMDbPY==6.8
lxml==4.5.0
Pillow==7.0.0
pkg-resources==0.0.0
pytz==2019.3
soupsieve==1.9.5
SQLAlchemy==1.3.13
sqlparse==0.3.0
